Turned plugin into an external module.
